Refactor stub methods
